refactor: 피드백 반영, RoundDao 인터페이스 추출
